Lower the matrix determinant epsilon
The determinant of the matrix is proportional to the volume of the
tetrahedron formed by the edges of the triangle. When viewed almost
horizontally, this would be smaller than the previous epsilon value,
though still positive. Lowering the epsilon fixed this case.

Unfortunately, some small faces are still dropped. Avoid them...?