Separate the handling of depth and distortion
In a perspective projection, depth and the degree of distortion are
the same, or at least proportional. This allowed us to use the
distortion value w in the rasterizer as a depth value, and ignore
the z coordinate entirely. Interpolating the true depth is more
correct and also allows us to use orthographic projections.

Also corrected aspect ratio in the demo.