Expand sample, add a simpler sample, and patch bug
Added ground and a backdrop to the main sample. Also created
another sample, containing just one triangle, in order to simplify
testing.

The ground exposed a bug in the rasterizer: when a triangle passed
through the camera plane (z=0), the triangle bounding box was
incorrectly calculated. I've just patched it for now, but performance
has taken a real hit.