Improve bound calculation for faces with nonpositive w-coordinates
Previously, in order to avoid incorrect bounds calculation for
faces with vertices with at least one nonpositive w-coordinate, we
checked each pixel in the framebuffer, which was very expensive.

With the new method, if a vertex has a negative or zero w, we fully
extend the bounds in the direction of the x and y coordinates (that
is, any positive x-value extends the x-max of the bounds to 1, any
negative or zero x-value extends the x-min to -1, and the same for
y). This is not optimal, but it's fairly simple and should be 2-3
times better on average.

The bounds-checking function also checks for faces with no positive
w-coordinates and drops them, as well as faces outside the frame.

The scene sample now renders at ~190 FPS instead of 46!